feat: add support for external event id
This allows us to pair ScheduledEvents on ATSes with Events on GoodTime with certainty.
Otherwise, we will pair ScheduledEvents on ATSes with Events on GoodTime based on other properties (like start & end time) which might not be super accurate since GoodTime allows creating two Events in the same Interview that start at the same time.